+++ saleor/graphql/attribute/mutations/attribute_value_delete.py
@@ -2,7 +2,6 @@ import graphene
 from django.db.models import Exists, OuterRef, Q
 
 from ....attribute import models as models
-from ....permission.enums import ProductTypePermissions
 from ....product import models as product_models
 from ...core import ResolveInfo
 from ...core.descriptions import ADDED_IN_310
@@ -10,6 +9,7 @@ from ...core.mutations import ModelDeleteMutation, ModelWithExtRefMutation
 from ...core.types import AttributeError
 from ...plugins.dataloaders import get_plugin_manager_promise
 from ..types import Attribute, AttributeValue
+from ..utils import check_permissions_for_attribute
 
 
 class AttributeValueDelete(ModelDeleteMutation, ModelWithExtRefMutation):
@@ -23,10 +23,16 @@ class AttributeValueDelete(ModelDeleteMutation, ModelWithExtRefMutation):
         )
 
     class Meta:
+        auto_permission_message = False
         model = models.AttributeValue
         object_type = AttributeValue
-        description = "Deletes a value of an attribute."
-        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        description = (
+            "Deletes a value of an attribute.\n\n"
+            "Depending on the attribute type, it requires different permissions to delete:\n"
+            "`PRODUCT_TYPE` requires `MANAGE_PRODUCTS` or `MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES` permissions,\n"
+            "`PAGE_TYPE` requires `MANAGE_PAGES` or `MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES` permissions.\n"
+            "DEPRECATED: it will be changed in 3.15."
+        )
         error_type_class = AttributeError
         error_type_field = "attribute_errors"
 
@@ -35,6 +41,7 @@ class AttributeValueDelete(ModelDeleteMutation, ModelWithExtRefMutation):
         cls, _root, info: ResolveInfo, /, *, external_reference=None, id=None
     ):
         instance = cls.get_instance(info, external_reference=external_reference, id=id)
+        check_permissions_for_attribute(info.context, instance.attribute)
         product_ids = cls.get_product_ids_to_update(instance)
         response = super().perform_mutation(
             _root, info, external_reference=external_reference, id=id